Update `CreatePersonExpander`
Fix the initial `nameError`, which should reflect the initial value of
`name` rather than being hard coded (to an incorrect no-error string).

The mechanism to suppress the error message on the first opening is
`showNameError`, not an incorrect `nameError` value. The value of
`nameError` must always be correct, e.g. to disable the create button.