Extract people API into a separate controller
Using a separate controller with `ApiController` attribute simplifies
returning a 400 response for model binding errors.

Front end code for distinguishing between a model binding error and
an application validation error had to be updated as well because
a non empty body is returned from the server for both types of errors
and we cannot just check if the body is empty.