Update delete person endpoint
Specify the name of the person to delete in the route. The main reason
for this is because in order to model bind a controller's action string
parameter to a simple request body text we need either
1. send the text as a JSON string (with quotes and escaping), not
a raw text or
2. configure the front end to set the content-type request header to
text/plain.
Using a route parameter requires just adding `FromRoute` attribute and
everyone is happy.

Additionally, this approach seems more REST-like: a route with the name
is a URI of the resource to DELETE.