Update 404 policy
Return an empty body when the action not found and an empty object
when the entity not found. This seems to be more logical than the
previous policy of returning a string when the action not found and an
empty body when the entity not found because returning an empty body
for an unrecognized action discloses nothing about the routes structure
of the app.