Update patterns / predicates
Rename `patterns.js` into `predicates.js` as they all are just
functions.

Abandon the idea of the nested `isObject` and `isArray` predicates.
This is beautiful but less readable and requires more care to make it
work properly. So, instead of a predicate "this is an object with
a field 'name' of type string" we will use a more flattened predicate
"this thing is an object and this thing has a property with name 'name'
and type string.